backend/requirements.txt
fastapi
uvicorn
sse-starlette
langchain-google-genai
langgraph>=0.1.7
pydantic
python-dotenv
yfinance
beautifulsoup4
requests
pandas
langchain>=0.2.16
langchain-core>=0.2.39
langchain-openai>=0.1.20
faiss-cpu>=1.8.0.post1
numpy>=1.26.4
sentence-transformers>=3.1.1
duckduckgo-search>=6.3.2
praw>=7.7.1
